+++ plugins/postcss-nesting/src/lib/atrule-within-rule.ts
@@ -1,10 +1,13 @@
 import cleanupParent from './cleanup-parent.js';
+import { options } from './options.js';
 import shiftNodesBeforeParent from './shift-nodes-before-parent.js';
 import validAtrules from './valid-atrules.js';
+import { walkFunc } from './walk-func.js';
+import type { AtRule, Rule } from 'postcss';
 
-export default function atruleWithinRule(node, walk, opts) {
+export default function atruleWithinRule(node: AtRule, parent: Rule, walk: walkFunc, opts: options) {
 	// move previous siblings and the node to before the parent
-	const parent = shiftNodesBeforeParent(node);
+	shiftNodesBeforeParent(node, parent);
 
 	// clone the parent as a new rule with children appended to it
 	const rule = parent.clone().removeAll().append(node.nodes);
@@ -19,4 +22,6 @@ export default function atruleWithinRule(node, walk, opts) {
 	walk(rule, opts);
 }
 
-export const isAtruleWithinRule = (node) => node.type === 'atrule' && validAtrules.includes(node.name) && Object(node.parent).type === 'rule';
+export function isAtruleWithinRule(node: AtRule) {
+	return validAtrules.includes(node.name);
+}