+++ plugins/postcss-nesting/src/lib/rule-within-rule.ts
@@ -1,16 +1,18 @@
 import shiftNodesBeforeParent from './shift-nodes-before-parent.js';
 import cleanupParent from './cleanup-parent.js';
 import mergeSelectors from './merge-selectors/merge-selectors.js';
+import type { Rule } from 'postcss';
+import { options } from './options.js';
 
-export default function transformRuleWithinRule(node, opts) {
+export default function transformRuleWithinRule(node: Rule, parent: Rule, opts: options) {
 	// move previous siblings and the node to before the parent
-	const parent = shiftNodesBeforeParent(node);
+	shiftNodesBeforeParent(node, parent);
 
 	// update the selectors of the node to be merged with the parent
 	node.selectors = mergeSelectors(parent.selectors, node.selectors, opts);
 
 	// merge similar rules back together
-	const areSameRule = (node.type === 'rule' && parent.type === 'rule' && node.selector === parent.selector) || (node.type === 'atrule' && parent.type === 'atrule' && node.params === parent.params);
+	const areSameRule = (node.type === 'rule' && parent.type === 'rule' && node.selector === parent.selector);
 
 	if (areSameRule) {
 		node.append(...parent.nodes);
@@ -20,4 +22,9 @@ export default function transformRuleWithinRule(node, opts) {
 	cleanupParent(parent);
 }
 
-export const isRuleWithinRule = (node) => node.type === 'rule' && Object(node.parent).type === 'rule' && node.selectors.every((selector) => selector.trim().indexOf('&') === 0 && selector.indexOf('|') === -1);
+export function isValidRuleWithinRule(node: Rule) {
+	return node.selectors.every((selector) => {
+		return selector.trim().indexOf('&') === 0 &&
+				selector.indexOf('|') === -1;
+	});
+}