+++ plugins/postcss-nesting/src/lib/shift-nodes-before-parent.ts
@@ -1,7 +1,7 @@
+import type { ChildNode, Container } from 'postcss';
 import cleanupParent from './cleanup-parent';
 
-export default function shiftNodesBeforeParent(node) {
-	const parent = node.parent;
+export default function shiftNodesBeforeParent(node: ChildNode, parent: Container<ChildNode>) {
 	const index = parent.index(node);
 
 	// conditionally move previous siblings into a clone of the parent
@@ -14,6 +14,4 @@ export default function shiftNodesBeforeParent(node) {
 	// move the current node before the parent (and after the conditional clone)
 	parent.before(node);
 	parent.raws.semicolon = true; /* nested rules end with "}" and do not have this flag set */
-
-	return parent;
 }