+++ plugins/postcss-cascade-layers/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - Updated: Support for Node v14+ (major).
 
+### Unreleased
+
+- Add support for `@scope` and `@container` as parent rules for `@layer`
+
 ### 1.0.6 (September 7, 2022)
 
 - Fix broken `@keyframes` in `@layer`.